+++ src/default_ports.js
@@ -1,0 +1,8 @@
+module.exports = Object.freeze({
+    socks: 9050,
+    http: 9080,
+    dns: 9053,
+    control: 9077,
+    controlWs: 9078,
+    default_host: null
+});